Migrate to separate module rpcinfo
