+++ components/analytics/shake2report/api/build.gradle.kts
@@ -6,5 +6,4 @@ dependencies {
     implementation(libs.cicerone)
 
     implementation(projects.components.bridge.api)
-    implementation(projects.components.info.api)
 }
