+++ components/analytics/shake2report/api/src/main/java/com/flipperdevices/shake2report/api/Shake2ReportApi.kt
@@ -2,8 +2,6 @@ package com.flipperdevices.shake2report.api
 
 import android.content.Context
 import com.flipperdevices.bridge.api.model.FlipperGATTInformation
-import com.flipperdevices.info.api.model.FlipperRpcInformation
-import com.flipperdevices.info.api.model.FlipperStorageInformation
 import com.github.terrakok.cicerone.Screen
 
 interface Shake2ReportApi {
@@ -15,9 +13,9 @@ interface Shake2ReportApi {
 
     fun reportBugScreen(context: Context): Screen?
 
+
     fun updateGattInformation(gattInformation: FlipperGATTInformation)
-    fun updateRpcInformation(rpcInformation: FlipperRpcInformation)
-    fun updateStorageInformation(storageInfo: FlipperStorageInformation)
+    fun setExtra(tags: List<Pair<String, String>>)
 
     fun reportException(
         throwable: Throwable,
