+++ components/analytics/shake2report/impl/src/main/java/com/flipperdevices/analytics/shake2report/impl/api/Shake2ReportApiImpl.kt
@@ -4,11 +4,8 @@ import android.content.Context
 import android.content.Intent
 import com.flipperdevices.analytics.shake2report.impl.InternalShake2Report
 import com.flipperdevices.analytics.shake2report.impl.activity.Shake2ReportActivity
-import com.flipperdevices.analytics.shake2report.impl.helper.FlipperInformationMapping
 import com.flipperdevices.bridge.api.model.FlipperGATTInformation
 import com.flipperdevices.core.di.AppGraph
-import com.flipperdevices.info.api.model.FlipperRpcInformation
-import com.flipperdevices.info.api.model.FlipperStorageInformation
 import com.flipperdevices.shake2report.api.Shake2ReportApi
 import com.github.terrakok.cicerone.Screen
 import com.github.terrakok.cicerone.androidx.ActivityScreen
@@ -30,15 +27,11 @@ class Shake2ReportApiImpl @Inject constructor(
     }
 
     override fun updateGattInformation(gattInformation: FlipperGATTInformation) {
-        internalShake2Report.setExtra(FlipperInformationMapping.convert(gattInformation))
+        internalShake2Report.setExtra(convert(gattInformation))
     }
 
-    override fun updateRpcInformation(rpcInformation: FlipperRpcInformation) {
-        internalShake2Report.setExtra(FlipperInformationMapping.convert(rpcInformation))
-    }
-
-    override fun updateStorageInformation(storageInfo: FlipperStorageInformation) {
-        internalShake2Report.setExtra(FlipperInformationMapping.convert(storageInfo))
+    override fun setExtra(tags: List<Pair<String, String>>) {
+        internalShake2Report.setExtra(tags)
     }
 
     override fun reportException(throwable: Throwable, tag: String?, extras: Map<String, String>?) {
@@ -48,4 +41,29 @@ class Shake2ReportApiImpl @Inject constructor(
 
         Sentry.captureEvent(event)
     }
+}
+
+private const val FLIPPER_DEVICE_NAME = "flipper_device_name"
+private const val FLIPPER_MANUFACTURER_NAME = "flipper_manufacturer_name"
+private const val FLIPPER_HARDWARE_REVISION = "flipper_hardware_revision"
+private const val FLIPPER_SOFTWARE_REVISION = "flipper_software_revision"
+private const val FLIPPER_BATTERY_LEVEL = "flipper_battery_level"
+private fun convert(gattInformation: FlipperGATTInformation): List<Pair<String, String>> {
+    val tagsList = mutableListOf<Pair<String, String>>()
+    gattInformation.deviceName?.let {
+        tagsList.add(FLIPPER_DEVICE_NAME to it)
+    }
+    gattInformation.manufacturerName?.let {
+        tagsList.add(FLIPPER_MANUFACTURER_NAME to it)
+    }
+    gattInformation.hardwareRevision?.let {
+        tagsList.add(FLIPPER_HARDWARE_REVISION to it)
+    }
+    gattInformation.softwareVersion?.let {
+        tagsList.add(FLIPPER_SOFTWARE_REVISION to it)
+    }
+    gattInformation.batteryLevel?.let {
+        tagsList.add(FLIPPER_BATTERY_LEVEL to it.toString())
+    }
+    return tagsList
 }
