+++ components/analytics/shake2report/noop/build.gradle.kts
@@ -8,7 +8,6 @@ dependencies {
     implementation(projects.components.analytics.shake2report.api)
     implementation(projects.components.bridge.api)
     implementation(projects.components.core.di)
-    implementation(projects.components.info.api)
 
     implementation(libs.cicerone)
     implementation(libs.annotations)
