+++ components/analytics/shake2report/noop/src/main/java/com/flipperdevices/shake2report/noop/Shake2ReportStub.kt
@@ -3,8 +3,6 @@ package com.flipperdevices.shake2report.noop
 import android.content.Context
 import com.flipperdevices.bridge.api.model.FlipperGATTInformation
 import com.flipperdevices.core.di.AppGraph
-import com.flipperdevices.info.api.model.FlipperRpcInformation
-import com.flipperdevices.info.api.model.FlipperStorageInformation
 import com.flipperdevices.shake2report.api.Shake2ReportApi
 import com.github.terrakok.cicerone.Screen
 import com.squareup.anvil.annotations.ContributesBinding
@@ -22,11 +20,7 @@ class Shake2ReportStub @Inject constructor() : Shake2ReportApi {
         // Do nothing
     }
 
-    override fun updateRpcInformation(rpcInformation: FlipperRpcInformation) {
-        // Do nothing
-    }
-
-    override fun updateStorageInformation(storageInfo: FlipperStorageInformation) {
+    override fun setExtra(tags: List<Pair<String, String>>) {
         // Do nothing
     }
