+++ components/bridge/rpcinfo/api/src/main/java/com/flipperdevices/bridge/rpcinfo/model/FlipperInformationStatus.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.info.api.model
+package com.flipperdevices.bridge.rpcinfo.model
 
 sealed class FlipperInformationStatus<T> {
     class NotStarted<T> : FlipperInformationStatus<T>()
