+++ components/bridge/rpcinfo/api/src/main/java/com/flipperdevices/bridge/rpcinfo/model/FlipperRpcInformation.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.info.api.model
+package com.flipperdevices.bridge.rpcinfo.model
 
 import com.flipperdevices.core.data.SemVer
 import kotlinx.collections.immutable.ImmutableMap
