+++ components/bridge/rpcinfo/impl/build.gradle.kts
@@ -1,0 +1,29 @@
+plugins {
+    id("flipper.android-lib")
+    id("com.squareup.anvil")
+    id("kotlin-kapt")
+}
+
+dependencies {
+    implementation(projects.components.bridge.rpcinfo.api)
+
+    implementation(projects.components.core.di)
+    implementation(projects.components.core.log)
+    implementation(projects.components.core.ktx)
+    implementation(projects.components.core.data)
+
+    implementation(projects.components.bridge.api)
+    implementation(projects.components.bridge.service.api)
+    implementation(projects.components.bridge.pbutils)
+
+    implementation(projects.components.analytics.shake2report.api)
+    implementation(projects.components.analytics.metric.api)
+
+    implementation(libs.kotlin.immutable.collections)
+    implementation(libs.kotlin.coroutines)
+
+    // Dagger deps
+    implementation(libs.dagger)
+    kapt(libs.dagger.kapt)
+
+}
