+++ components/bridge/rpcinfo/impl/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.flipperdevices.bridge.rpcinfo.impl" />
