+++ components/bridge/rpcinfo/impl/src/main/java/com/flipperdevices/bridge/rpcinfo/impl/FlipperRpcInformationApiImpl.kt
@@ -1,39 +1,29 @@
-package com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers
+package com.flipperdevices.bridge.rpcinfo.impl
 
 import com.flipperdevices.bridge.api.utils.Constants
+import com.flipperdevices.bridge.rpcinfo.api.FlipperRpcInformationApi
+import com.flipperdevices.bridge.rpcinfo.impl.fullinforpc.DeprecatedFlipperFullInfoRpcApi
+import com.flipperdevices.bridge.rpcinfo.impl.fullinforpc.NewFlipperFullInfoRpcApi
+import com.flipperdevices.bridge.rpcinfo.impl.shaketoreport.FlipperInformationMapping
+import com.flipperdevices.bridge.rpcinfo.model.FlipperInformationStatus
+import com.flipperdevices.bridge.rpcinfo.model.FlipperRpcInformation
 import com.flipperdevices.bridge.service.api.FlipperServiceApi
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ktx.jre.withLock
 import com.flipperdevices.core.log.LogTagProvider
-import com.flipperdevices.info.api.model.FlipperInformationStatus
-import com.flipperdevices.info.api.model.FlipperRpcInformation
-import com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers.fullinforpc.DeprecatedFlipperFullInfoRpcApi
-import com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers.fullinforpc.NewFlipperFullInfoRpcApi
 import com.flipperdevices.shake2report.api.Shake2ReportApi
 import com.squareup.anvil.annotations.ContributesBinding
+import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
-import javax.inject.Inject
-
-interface FlipperRpcInformationApi {
-    fun getRpcInformationFlow(): StateFlow<FlipperInformationStatus<FlipperRpcInformation>>
-    suspend fun invalidate(
-        scope: CoroutineScope,
-        serviceApi: FlipperServiceApi,
-        force: Boolean = false
-    )
-
-    suspend fun reset()
-}
 
 @ContributesBinding(AppGraph::class, FlipperRpcInformationApi::class)
 class FlipperRpcInformationApiImpl @Inject constructor(
@@ -95,7 +85,7 @@ class FlipperRpcInformationApiImpl @Inject constructor(
                         it
                     }
                 }
-                shake2ReportApi.updateRpcInformation(rpcInformation)
+                shake2ReportApi.setExtra(FlipperInformationMapping.convert(rpcInformation))
             }
 
         rpcInformationFlow.update {
