+++ components/bridge/rpcinfo/impl/src/main/java/com/flipperdevices/bridge/rpcinfo/impl/fullinforpc/NewFlipperFullInfoRpcApi.kt
@@ -1,8 +1,8 @@
-package com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers.fullinforpc
+package com.flipperdevices.bridge.rpcinfo.impl.fullinforpc
 
 import com.flipperdevices.bridge.api.manager.FlipperRequestApi
 import com.flipperdevices.bridge.api.model.wrapToRequest
-import com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers.mapper.NewFlipperRpcInfoMapper
+import com.flipperdevices.bridge.rpcinfo.impl.mapper.NewFlipperRpcInfoMapper
 import com.flipperdevices.protobuf.main
 import com.flipperdevices.protobuf.property.getRequest
 import kotlinx.coroutines.flow.merge
