+++ components/bridge/rpcinfo/impl/src/main/java/com/flipperdevices/bridge/rpcinfo/impl/mapper/FlipperRpcInfoMapper.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.bridge.rpcinfo.impl.mapper
+
+import com.flipperdevices.bridge.rpcinfo.model.FlipperRpcInformation
+
+internal interface FlipperRpcInfoMapper {
+    fun map(raw: InternalFlipperRpcInformationRaw): FlipperRpcInformation
+}
