+++ components/bridge/rpcinfo/impl/src/main/java/com/flipperdevices/bridge/rpcinfo/impl/mapper/NewFlipperRpcInfoMapper.kt
@@ -1,9 +1,9 @@
-package com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers.mapper
+package com.flipperdevices.bridge.rpcinfo.impl.mapper
 
-import com.flipperdevices.info.api.model.FirmwareInfo
-import com.flipperdevices.info.api.model.FlipperDeviceInfo
-import com.flipperdevices.info.api.model.FlipperRpcInformation
-import com.flipperdevices.info.api.model.RadioStackInfo
+import com.flipperdevices.bridge.rpcinfo.model.FirmwareInfo
+import com.flipperdevices.bridge.rpcinfo.model.FlipperDeviceInfo
+import com.flipperdevices.bridge.rpcinfo.model.FlipperRpcInformation
+import com.flipperdevices.bridge.rpcinfo.model.RadioStackInfo
 import kotlinx.collections.immutable.toImmutableMap
 
 private const val DEVICE_NAME = "hardware.name"
