+++ components/info/api/build.gradle.kts
@@ -3,11 +3,7 @@ plugins {
 }
 
 dependencies {
-    implementation(libs.cicerone)
     implementation(projects.components.deeplink.api)
 
-    implementation(projects.components.core.data)
     implementation(projects.components.core.ui.navigation)
-
-    implementation(libs.kotlin.immutable.collections)
 }
