+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/model/FlipperBasicInfo.kt
@@ -1,8 +1,8 @@
 package com.flipperdevices.info.impl.model
 
 import androidx.compose.runtime.Stable
-import com.flipperdevices.info.api.model.FlipperInformationStatus
-import com.flipperdevices.info.api.model.FlipperStorageInformation
+import com.flipperdevices.bridge.rpcinfo.model.FlipperInformationStatus
+import com.flipperdevices.bridge.rpcinfo.model.FlipperStorageInformation
 import com.flipperdevices.updater.model.FirmwareVersion
 
 @Stable
