--- components/info/impl/src/main/java/com/flipperdevices/info/impl/viewmodel/deviceinfo/helpers/mapper/FlipperRpcInfoMapper.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.info.impl.viewmodel.deviceinfo.helpers.mapper
-
-import com.flipperdevices.info.api.model.FlipperRpcInformation
-
-internal interface FlipperRpcInfoMapper {
-    fun map(raw: InternalFlipperRpcInformationRaw): FlipperRpcInformation
-}
